When Aodh alarm is deleted, need to update its state to INACTIVE
When Aodh alarm is deleted, need to update its state to INACTIVE.
This is important, as we do not really delete the alarm from the
graph, rather we keep it with is_deleted=t. We don't want it to
be returned by mistake by a query that filters active alarms.

Change-Id: Ide62f07607181eff1e6f55f907f10eaaadc3095a
Fixes: bug #1586747
Signed-off-by: dongwenjuan <dong.wenjuan@zte.com.cn>