ReScheduler class for Vitrage Snapshot Service
The ReScheduler is a decorating class to Python's
sched package Scheduler. Allows scheduling tasks for a
repeated number of times while providing a mechanism
to reschedule differently in case of a task's
failure. The class is now used to reschedule repeatedly
Vitrage's datasources get_all. This allows for better
get all exception\fault management as well as more
options for timing them.

Change-Id: I3f6131a7e1d2750405ddeed5575976d5a486a08d