delete 'resource_id' as a component of key_values
As 'alarm_id' is unique and used as key of vitrage_id, and meanwhile
Aodh alarm notifications may not include the 'resource_id' in it's
payload(only alarm.creation notification include the 'resource_id).
If we still use 'resource_id' as a component of key_value, it will
caused different 'vitrage_id' with the same alarm in Aodh transform.

Signed-off-by: dongwenjuan <dong.wenjuan@zte.com.cn>

Change-Id: Id71c0d8b483a67135f16ecfb64c2d453ce667bc5