Bug fix: When creating a vertex by the evaluator transformer, take only parameters that are not None. Otherwise, a None parameter may overwrite a "real" parameter that already exists on the vertex.
An example:
In case there are two actions on a host: set-state to ERROR and mark-down.
The first vertex update is: vitrage_state=ERROR, is_marked_down=None
The second vertex update is: vitrage_state=None, is_marked_down=True

The second vertex update replaces the ERROR state to with None.
My fix removes the None, so the ERROR state remains as is.

Change-Id: If523bb9c06bcd3bd289f4de956fe737fc573a2da
Closes-Bug: #1645659