for ML - add resource details: id and type to alarms
this was done by adding type and category to alarms neighbors
in raise_alarm and evaluator_event_transformer,
and using it later in the processor.
the tests also was modified accordingly.

those detailes will be needed later to recognize alarm uniquely
in order to compute the total time an alarms was active.

Change-Id: Id6db43de1595d2bf73db5a4db83472e6c2a14a8e