Aligning install guide with the Pike doc structure
In Pike OpenStack documentation is changed in a way that the project repos
contain most of the documentation material [1]. This affcects how the doc
folder shall be strucutred inside the project repo. With this change I
align the install guide to the new structure and correct the warnings
during the document generation. The overall work is tracked in [2].

[1]: http://specs.openstack.org/openstack/docs-specs/specs/pike/os-manuals-migration.html
[2]: https://etherpad.openstack.org/p/doc-migration-tracking

Change-Id: Ib5400247d2ba5f8dac80ef68a43d06868bd42f1c
Depends-On: Ia750cb049c0f53a234ea70ce1f2bbbb7a2aa9454
Signed-off-by: csatari <gergely.csatari@nokia.com>