Refactor the Vitrage template content validators and their tests.
The motivations is that adding a new validator (e.g. for the new execute-mistral action) will be done in a separate and independent file.
As part of the refactoring, I moved the validation of the existence of the 'action_target' block from the syntax validator to the content validator, since in the execute-mistral action it won't be a mandatory field.

In the next change I will move template_content_validator and test_template_content_validator under the relevant 'content' packages. I removed it from this change since it caused git problems.

Change-Id: I97fe6d05e214e746f6129e3d18f805e830e21cf7
Implements: blueprint support-external-actions