Fix a bug in the profiler serializer.
The profiler serializer, which is used if the osprofiler is installed,
was initialized with None and therefore did not serialize the context.
As a result, the context did not reach the api handler layer and was
ignored.

The solution: if the serializer is None, return the un-serialized
context as is.

Change-Id: I2317cff4014db2b4e437b2e9428bd936518d7ba8
Closes-Bug: #1707589