Added a version to Vitrage templates
Added a new "version" field to the metadata section of Vitrage templates, in order to support future enhancements that are not backward-compatible. The default version is 1.
Version 2 will be added in anther change (with execute-mistral changes), together with a mechanism for running the correct validator and loader per-version.

Change-Id: I25a7bf50a2ab7277f5bbdc551bf676c62dfd6fdc
Implements: blueprint refactor-execute-mistral-definition