Refactor template loading
The template loading code was extracted from the TemplateData class and
placed in separate files, to allow future support of per-version loading
(i.e. different loaders for different template versions).
No real change was done in the code, it  was just split to separate files.

The changes:
* Separate the loading logic from the data
* Split the template loading to specific loaders: template, scenario and equivalences

Change-Id: I810da2682f3c1804312b26041b73280c040432a5
Implements: blueprint refactor-execute-mistral-definition