refactoring vitrage processes
oslo service replaced by cotyledon.
Motivation:
Oslo service uses eventlet and green threads for timers and threads.
In Vitrage we use real threads and multiprocessing, it seems this 
combination causes unexpected behavior (such as timers not running, 
processes not starting) so the entire solution is very delicate to 
changes.

As part of an ongoing effort (previously, monkeypatching was removed)
this commit solves issues encountered in rpc_collector commit in 
addition it removes unnecessary complications in the application.

Co-Authored-By: Eyal <eyal.bar-ilan@nokia.com>
Change-Id: If33708e128d7aeb420a9aa79f71cfccd2b48dfb3