Delete outdated entities by the consistency
As part of Rocky fast-failover support,
vitrage-graph is now reloaded from the
database. This causes an issue with
datasources using caches, that can become
outdated in two cases

- After vitrage-graph restart. This is
handled by the consistency service.

- If more than one vitrage-collector is used.

The solution/workaround for Rocky is to let
the consistency enforcer delete outdated
entities for datasources that specifically
request it.

For Stein, we should move this handling to
the processor (so the entities will be
deleted immediately) and delete all the
caches

Change-Id: I953137dc870f48ed42acba93789bb947809c41cc