bugfix: update a placeholder vertex can add properties to a real vertex
Currently, if a placeholder vertex is created and the real vertex
already exists in the graph, the placeholder vertex is ignored.
With this change, new properties that are part of the placeholder
vertex will be added to the existing real vertex.

This fix is similar to the one done in Queens:
https://review.openstack.org/#/c/566832

With small changes due to Networkx 2.0 support

Change-Id: Idf0383fa08ddb6936868e74b7ea8e727a9935523